chore(core): move BaseCrossEncoder to langchain-core (#35809)
