+++ libs/langchain/langchain_classic/retrievers/document_compressors/cross_encoder.py
@@ -1,16 +1,3 @@
-from abc import ABC, abstractmethod
+from langchain_core.cross_encoders import BaseCrossEncoder
 
-
-class BaseCrossEncoder(ABC):
-    """Interface for cross encoder models."""
-
-    @abstractmethod
-    def score(self, text_pairs: list[tuple[str, str]]) -> list[float]:
-        """Score pairs' similarity.
-
-        Args:
-            text_pairs: List of pairs of texts.
-
-        Returns:
-            List of scores.
-        """
+__all__ = ["BaseCrossEncoder"]
